Hi Adam. Here are some notes on this commit:
How it works

To use the 'custom' wave shpe, you pass in either a waveShape, or a waveTable parameter into the bap.oscillator constructor. A waveShape is a non-negative array of floating point values. If any number in waveShape is negative, it'll throw an error.

waveTable should look like this:

var waveTable = {
  real: [0, 0.1, 0.2, ....],
  imag: [0, 0.1, 0.2, ....],
};

where real, and imag are the real and imaginary parts to the frequency numbers of Fourier decompositions. If the real and imag arrays don't match in length, an error is thrown.

I modified the metronome example to use the custom wave shape. It should work off the bat. Comment out either the waveShape or the waveTable line, and it will work the same.

I can't create more than 6 AudioContexts, so I'm storing an AudioContext in the window, using a sort of Singleton pattern. Similarly, I'm "caching" DFT calculations of old waveShapes in a the same bap namespace in the window object. I have a lot to learn as a programmer, so if you know of a better way to do this, I'm more than grateful for any guidance.

Issues I ran into

* I had trouble with generating the examples' bundle.js file using 'npm run-script build:examples'. The error returned is "Cannot find module 'lodash.debounce'". What I had to do to get around this was copy the bundle.js file that you're using in the online examples into the examples folder, and link to it in index.html. I then modified bundle.js file directly. I've tried to edit the original js files (e.g. Oscillator.js, metronome.js, etc.) that the bundle.js is generated from, but I'm not 100% positive it'll work once you correctly build it.

* I had to comment out the default argument of "sine" in the Oscillator's mixin, because for some reason, it always over-wrote whatever oscillator shape you passed into it